refactor(simulation): remove create-detail web cycle
